js/datachannel_test_2peer/id/009.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - ID Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
            var label = "test";
            var dataChannel = null;
            var dataChannelOptions = {
                id : 6
            };
            var test = async_test("Set up a DataChannels with a specific id: " + dataChannelOptions.id + " - close the Channel - set up another DataChannel with the same id (reuse id)");
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label, dataChannelOptions);
                    dataChannel.onopen = test.step_func(function() {
                        assert_equals(dataChannel.id, dataChannelOptions.id, "ID not set correct: ");
                        dataChannel.close();
                        setTimeout(test.step_func(function() {
                        	try{
                        		dataChannel = peerConnection.createDataChannel(label, dataChannelOptions);
		                        dataChannel.onopen = test.step_func(function() {
		                            assert_equals(dataChannel.id, dataChannelOptions.id, "ID not set correct: ");
		                            test.done();
		                        });
                        	}catch(e){
                        		assert_unreached(e.name + ": " + e.message);
                        	}
	                    }), 5000);
                    });
                });
            } else {
            	var i = 0;
                peerConnection.ondatachannel = test.step_func(function(e) {
                    i++;
                    dataChannel = e.channel;
       				dataChannel.onopen = test.step_func(function(){
       					if(i ==2){
	                    	assert_equals(dataChannel.id, dataChannelOptions.id, "ID not set correct: ");
	                    	test.done();
	                    }
       				});

                });
            }
		</script>
	</body>
</html>
